REBASE: Stripped unedited MCP files from repository.
Two commits have been stripped from the repository history, effectively
discarding all of the MCP files not specific to the project.  This creates a
much longer "Untracked files" list when injected into MCP, but also makes the
repository itself much cleaner.

Because the history has been mucked with, updating to this commit is a pain.
The instructions here are directed at command-line git.  For anything else,
they may be useful, but you're on your own; I'm at least as lost as you.

If you don't have any local changes to mcp_interface itself (your mods/ don't
count for this), do this from the mcp_interface directory:

1. Remove your current clone's brain.
   $ rm -r .git
2. Re-clone the repository.
   $ git clone --branch mcp_interface https://github.com/FunnyMan3595/mcp_deobfuscate.git gittemp
3. Steal the new clone's brain.
   $ mv gittemp/.git .
4. And bury the body.
   $ rm -rf gittemp/

If you do have local changes, it's a bit trickier:

1. Commit or stash any modifications.  Making a backup of the entire directory
   wouldn't be a bad idea, either, just in case.
2. Force-fetch the new version:
   $ git fetch -f origin mcp_interface
3. Start an interactive rebase onto the new version:
   $ git rebase -i origin/mcp_interface
4. Delete these two lines from the rebase list, then run it:
   pick f30b552 MCP 7.2
   pick 818a9c3 Forge modifications to MCP
5. During your rebase, it may fail to apply changes if you edited a file that
   isn't in the new set.  Just add the file and run 'git rebase --continue'.
6. Once the rebase is complete, you'll need to recreate the other MCP files.
   You could just copy in a fresh MCP install, but you can also do it with git:
   $ git checkout 818a9c3 # Go back to the commit we just removed from history.
   $ git rm --cached -r . # Tell git to leave everything we removed before.
   $ git checkout -f yourbranch # And then clobber what wasn't removed.